requirements.txt
Mezzanine>=3.1.10
Django<1.8
enum34>=1.0
po-localization>=0.5.3
django-bootstrap3>=4.11.0
django-braces>=1.4.0
django-import-export>=0.2.6
django-fsm>=2.2.0
django-extra-views>=0.6.5
django-utils2>=1.8.0
django-permission>=0.8.3
django-model-report==0.2.1
